+++ modules/rumors.php
@@ -442,7 +442,7 @@ if ($op=="rumor9"){
 }
 if ($op == "hof") {
 	page_header("Hall of Fame");
-	$page = httpget('page');
+	$page = (int) httpget('page');
 	$pp = 50;
 	$pageoffset = (int)$page;
 	if ($pageoffset > 0) $pageoffset--;
